requirements.txt
Django>=3.0.2,<3.1.0
djangorestframework>=3.11.0,<3.12.0
django-rest-auth>=0.9.5,<1.0.0
django-allauth>=0.41.0,<1.0.0

flake8>=3.7.9,<3.8.0
